+++ examples/resources/sdwan_cisco_secure_internet_gateway_feature_template/resource.tf
@@ -34,6 +34,7 @@ resource "sdwan_cisco_secure_internet_gateway_feature_template" "example" {
       ipsec_replay_window           = 1024
       ipsec_ciphersuite             = "aes256-cbc-sha1"
       ipsec_perfect_forward_secrecy = "group-14"
+      tracker                       = "test"
       track_enable                  = false
       tunnel_public_ip              = "5.5.5.5"
     }
@@ -60,6 +61,7 @@ resource "sdwan_cisco_secure_internet_gateway_feature_template" "example" {
       zscaler_surrogate_idle_time                     = 100
       zscaler_surrogate_display_time_unit             = "MINUTE"
       zscaler_surrogate_ip_enforce_for_known_browsers = true
+      zscaler_surrogate_refresh_time                  = 12345
       zscaler_surrogate_refresh_time_unit             = "MINUTE"
       zscaler_aup_enabled                             = true
       zscaler_aup_block_internet_until_accepted       = true
@@ -76,6 +78,7 @@ resource "sdwan_cisco_secure_internet_gateway_feature_template" "example" {
       name             = "TRACKER1"
       endpoint_api_url = "https://1.1.1.1"
       threshold        = 500
+      interval         = 60
       multiplier       = 4
       tracker_type     = "SIG"
     }