+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.7 (unreleased)
+
+- - Add `tracker`, `tracker_variable`, `interval`, `interval_variable` and `zscaler_surrogate_refresh_time` to `sdwan_cisco_secure_internet_gateway_feature_template` resource and data source
+
 ## 0.3.6
 
 - Fix issue with address family configuration of `sdwan_cisco_bgp_feature_template` resource